Generalize ks_stat to repeating data
Current implementation to get D_bottom subtracts (y - 1/len(data1)) from cdf. This assumes that the concave corner is just 1/n below the convex corner. This only happens when the data is unique. If there is repeating data, the concave corner could be multiples of 1/n below. 

y_shifted intermediate variable is added to find the concave corners. The concave corner of the current x is the convex corner of the previous x. Thus, i shifted the y values 1 index to the right, and added 0 on the left to fill up the space.

P.s i also switched the concave/convex terms in the comments in this file to be consistent with Datacamp's template and your verbal description in the video https://campus.datacamp.com/courses/case-studies-in-statistical-thinking/statistical-seismology-and-the-parkfield-region?ex=8.
 
The new implementation would find convex corners not just at 1/len(data1) below the concave corner, but also further below.
I can't find the random seed used in Datacamp, so I have tested this using my own np.random.seed(1) to get p = 0.246 for np.random.exponential with both the old and new implementation.